utils: Fix FileNotFoundError in temporary_directory()
Don't try to delete temporary directories that no longer exist.

Fixes: d22afa9d933529c01b74a6127fa41ce49bf0182b

+++ poetry/utils/helpers.py
@@ -48,7 +48,8 @@ def temporary_directory(*args, **kwargs):
 
     yield name
 
-    shutil.rmtree(name, onerror=_del_ro)
+    if os.path.exists(name):
+        shutil.rmtree(name, onerror=_del_ro)
 
 
 def get_cert(config, repository_name):  # type: (Config, str) -> Optional[Path]
